# Conflicts:
#	biz/adaptor/server.go
#	biz/application/service/file.go
#	biz/application/service/label.go
#	biz/application/service/user.go
#	biz/infrastructure/mapper/file/mongo.go
#	biz/infrastructure/mapper/sharefile/mongo.go
#	biz/infrastructure/mapper/user/es.go
#	biz/infrastructure/util/convertor/convertor.go
#	etc/config.yaml
#	go.mod
#	go.sum
#	provider/wire_gen.go

+++ biz/application/service/user.go
@@ -101,7 +101,7 @@ func (s *UserServiceImpl) CreateUser(ctx context.Context, req *gencontent.Create
 	if _, err = s.UserMongoMapper.Insert(ctx, &usermapper.User{
 		ID:          ID,
 		Name:        req.UserInfo.Name,
-		Sex:         req.UserInfo.Sex,
+		Sex:         int64(req.UserInfo.Sex),
 		Description: consts.DefaultDescription,
 		Url:         consts.DefaultAvatarUrl,
 	}); err != nil {
